+++ stack-it-up/common/src/main/java/com/chimericdream/stackitup/mixin/AccessorFurnaceInventory.java
@@ -1,0 +1,13 @@
+package com.chimericdream.stackitup.mixin;
+
+import net.minecraft.core.NonNullList;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.AbstractFurnaceBlockEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(AbstractFurnaceBlockEntity.class)
+public interface AccessorFurnaceInventory {
+    @Accessor
+    NonNullList<ItemStack> getItems();
+}