+++ stack-it-up/CHANGELOG.md
@@ -1,1 +1,6 @@
-### Unreleased
+### Unreleased changes
+
+#### New Features
+
+* Ported from the [AllStackable](https://www.curseforge.com/minecraft/mc-mods/all-stackable)/
+  AllStackableContinued fork onto Fabric and NeoForge for Minecraft 26.2.